# [Enhancement]Optimize publish version through batch in SHARED_DATA