--- a/be/src/service/service_be/starrocks_be.cpp
+++ b/be/src/service/service_be/starrocks_be.cpp
@@ -216,7 +216,7 @@ void start_be(const std::vector<StorePath>& paths, bool as_cn) {
     const auto service_name = "starrocks.lake.LakeService";
     const auto methods = {"abort_txn",           "abort_compaction", "compact",          "drop_table",
                           "delete_data",         "delete_tablet",    "get_tablet_stats", "publish_version",
-                          "publish_log_version", "vacuum",           "vacuum_full"};
+                          "publish_log_version", "publish_log_version_batch", "vacuum",   "vacuum_full"};
     for (auto method : methods) {
         brpc_server->MaxConcurrencyOf(service_name, method) = lake_service_max_concurrency;
     }