--- a/be/src/storage/lake/txn_log_applier.h
+++ b/be/src/storage/lake/txn_log_applier.h
@@ -18,8 +18,8 @@
 
 #include "common/status.h"
 #include "gutil/macros.h"
+#include "storage/lake/tablet_manager.h"
 #include "storage/lake/tablet_metadata.h"
-
 namespace starrocks::lake {
 
 class Tablet;