… be/src/storage/lake/tablet_metadata.h
… be/test/service/lake_service_test.cpp
… be/test/storage/lake/test_util.h
… fe/fe-core/src/main/java/com/starrocks/common/Config.java
… fe/fe-core/src/main/java/com/starrocks/journal/JournalEntity.java
… fe/fe-core/src/main/java/com/starrocks/lake/Utils.java
… fe/fe-core/src/main/java/com/starrocks/persist/EditLog.java
… fe/fe-core/src/main/java/com/starrocks/persist/OperationType.java
… fe/fe-core/src/main/java/com/starrocks/rpc/LakeService.java
… fe/fe-core/src/main/java/com/starrocks/transaction/DatabaseTransactionMgr.java
… fe/fe-core/src/main/java/com/starrocks/transaction/GlobalTransactionMgr.java
… fe/fe-core/src/main/java/com/starrocks/transaction/LakeTableTxnLogApplier.java
… fe/fe-core/src/main/java/com/starrocks/transaction/PublishVersionDaemon.java
… fe/fe-core/src/main/java/com/starrocks/transaction/TransactionGraph.java
… fe/fe-core/src/main/java/com/starrocks/transaction/TransactionStateBatch.java
… fe/fe-core/src/test/java/com/starrocks/catalog/FakeEditLog.java
… fe/fe-core/src/test/java/com/starrocks/catalog/GlobalStateMgrTestUtil.java
… fe/fe-core/src/test/java/com/starrocks/pseudocluster/PseudoBackend.java
… fe/fe-core/src/test/java/com/starrocks/transaction/DatabaseTransactionMgrTest.java
… fe/fe-core/src/test/java/com/starrocks/transaction/LakePublishBatchTest.java
… fe/fe-core/src/test/java/com/starrocks/transaction/TransactionGraphTest.java
… fe/fe-core/src/test/java/com/starrocks/transaction/TransactionStateBatchTest.java
… gensrc/proto/lake_service.proto