StarRocks adopts a new multi-version mechanism for storage and calculation separation, which requires transactions to be executed strictly serially in the final Publish Version stage (because the new Tablet Meta version needs to be generated based on the content of the previous version).
Since each Publish Version of Tablet are as follows:
1. Read the previous version of Tablet Meta
2. Read the Txn Log generated during this import stage
3. Write a new version of Tablet Meta

Due to the high latency characteristics of object storage, the latency of each Publish Version may be as high as hundreds of ms. Due to the strict serial execution characteristics of Publish Version, in high-concurrency real-time scenarios,  the performance bottleneck of load is stuck in the Publish Version stage. In addition, a large number of small files are generated, which is not friendly to object storage (especially privately deployed HDFS, etc.)

The pr performs a enhancemen through execution of the Publish Version task by batch , that is, multiple transactions can be processed in one publish task  to improve load throughput.

You can enable the feature by execute the sql
`ADMIN SET FRONTEND CONFIG ("enable_lake_batch_publish_version" = "true");`
and the max batch size can be adjusted by the config of `lake_batch_publish_max_version_num`

The performane of test is as follows,  
Duplicate  Key

|  Concurrency   | Time before optimization / s  |  Time after optimization / s | Improvement |
|  ----  | ----  | ---- | ---- |
| 50  | 64 | 11  | 5.8 |
| 100  | 134 | 22 | 6 |

Primayr Key
|  Concurrency   | Time before optimization / s  |  Time after optimization / s | Improvement |
|  ----  | ----  | ---- | ---- |
| 50  | 114 | 13  | 8.7 |
| 100  | 194 | 25 | 7.7 |

## What type of PR is this:

- [ ] BugFix
- [x] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [x] Yes, this PR will result in a change in behavior.
- [ ] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with different default values
- [x] Policy changes: use new policy to replace old one, functionality automatically enabled
- [ ] Feature removed
- [x] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [x] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function

## Bugfix cherry-pick branch check:

- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [x] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5