@wanpengfei-git 2023-08-21
#### [FE Incremental Coverage Report]

:disappointed: **fail** : 28 / 463 (06.05%)




#### file detail

|   |path|covered_line|new_line|coverage|not_covered_line_detail|
|----|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/transaction/LakeTableTxnLogApplier.java|0|38|00.00%|[94, 95, 96, 97, 99, 100, 101, 102, 103, 104, 105, 106, 107, 108, 110, 111, 112, 113, 115, 117, 118, 119, 120, 122, 123, 125, 126, 128, 129, 131, 132, 133, 134, 135, 136, 139, 140, 141]|
|:large_blue_circle:|com/starrocks/transaction/GlobalTransactionMgr.java|0|20|00.00%|[560, 561, 562, 563, 564, 565, 566, 567, 568, 570, 571, 606, 607, 608, 719, 720, 721, 722, 723, 724]|
|:large_blue_circle:|com/starrocks/transaction/TransactionStateBatch.java|0|74|00.00%|[37, 39, 43, 45, 48, 52, 53, 56, 60, 61, 64, 68, 69, 72, 74, 76, 77, 78, 81, 86, 87, 92, 93, 94, 95, 96, 97, 98, 99, 100, 103, 104, 105, 106, 111, 113, 114, 115, 116, 117, 120, 121, 125, 126, 128, 132, 136, 137, 139, 143, 147, 148, 150, 154, 158, 159, 162, 163, 166, 168, 169, 172, 177, 182, 183, 184, 185, 186, 189, 190, 191, 192, 194, 198]|
|:large_blue_circle:|com/starrocks/persist/EditLog.java|0|6|00.00%|[605, 606, 607, 608, 1598, 1599]|
|:large_blue_circle:|com/starrocks/transaction/PublishVersionDaemon.java|1|172|00.58%|[105, 106, 107, 108, 109, 110, 112, 187, 188, 190, 325, 327, 331, 332, 333, 334, 343, 344, 345, 346, 347, 348, 349, 350, 352, 353, 356, 374, 406, 408, 409, 411, 412, 414, 417, 418, 419, 420, 422, 423, 424, 425, 428, 429, 431, 432, 433, 434, 435, 436, 437, 438, 440, 441, 442, 443, 444, 446, 447, 448, 451, 452, 454, 458, 459, 460, 463, 464, 465, 466, 467, 468, 469, 470, 471, 472, 473, 474, 475, 478, 479, 482, 483, 484, 485, 486, 489, 492, 496, 497, 498, 502, 503, 504, 506, 508, 510, 513, 514, 515, 516, 518, 519, 521, 522, 524, 525, 527, 528, 530, 531, 533, 534, 535, 536, 538, 539, 540, 541, 545, 547, 548, 550, 551, 552, 553, 554, 555, 556, 558, 560, 562, 563, 565, 566, 567, 568, 569, 570, 573, 574, 575, 577, 578, 579, 581, 582, 584, 585, 586, 588, 589, 591, 593, 594, 595, 597, 598, 599, 601, 602, 603, 610, 612, 613, 614, 615, 616, 617, 684, 685]|
|:large_blue_circle:|com/starrocks/transaction/DatabaseTransactionMgr.java|2|94|02.13%|[722, 723, 725, 726, 727, 728, 730, 731, 733, 734, 735, 738, 739, 740, 741, 742, 744, 745, 747, 748, 751, 752, 753, 754, 757, 759, 762, 763, 764, 766, 768, 770, 1195, 1196, 1198, 1199, 1200, 1201, 1203, 1204, 1205, 1206, 1207, 1209, 1212, 1213, 1214, 1215, 1216, 1217, 1218, 1543, 1544, 1545, 1547, 1548, 1549, 1649, 1651, 1652, 1653, 1655, 1657, 1659, 1751, 1752, 1753, 1754, 1756, 1757, 1758, 1760, 1762, 1766, 1769, 1770, 1773, 1774, 1775, 1776, 1777, 1779, 1780, 1785, 1790, 1794, 1796, 1797, 1819, 1820, 1821, 1822]|
|:large_blue_circle:|com/starrocks/lake/Utils.java|3|13|23.08%|[168, 169, 170, 171, 172, 173, 196, 198, 199, 202]|
|:large_blue_circle:|com/starrocks/transaction/TransactionGraph.java|20|44|45.45%|[132, 195, 196, 198, 199, 200, 203, 204, 206, 207, 208, 209, 212, 213, 214, 215, 216, 219, 220, 221, 222, 223, 226, 227]|
|:large_blue_circle:|com/starrocks/common/Config.java|2|2|100.00%|[]|