@wanpengfei-git 2023-08-22
#### [FE Incremental Coverage Report]

:disappointed: **fail** : 23 / 476 (04.83%)




#### file detail

|   |path|covered_line|new_line|coverage|not_covered_line_detail|
|----|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/transaction/LakeTableTxnLogApplier.java|0|38|00.00%|[94, 95, 96, 97, 99, 100, 101, 102, 103, 104, 105, 106, 107, 108, 110, 111, 112, 113, 115, 117, 118, 119, 120, 122, 123, 125, 126, 128, 129, 131, 132, 133, 134, 135, 136, 139, 140, 141]|
|:large_blue_circle:|com/starrocks/transaction/GlobalTransactionMgr.java|0|20|00.00%|[560, 561, 562, 563, 564, 565, 566, 567, 568, 570, 571, 606, 607, 608, 719, 720, 721, 722, 723, 724]|
|:large_blue_circle:|com/starrocks/transaction/TransactionStateBatch.java|0|74|00.00%|[37, 39, 43, 45, 48, 52, 53, 56, 60, 61, 64, 68, 69, 72, 74, 76, 77, 78, 81, 86, 87, 92, 93, 94, 95, 96, 97, 98, 99, 100, 103, 104, 105, 106, 111, 113, 114, 115, 116, 117, 120, 121, 125, 126, 128, 132, 136, 137, 139, 143, 147, 148, 150, 154, 158, 159, 162, 163, 166, 168, 169, 172, 177, 182, 183, 184, 185, 186, 189, 190, 191, 192, 194, 198]|
|:large_blue_circle:|com/starrocks/persist/EditLog.java|0|6|00.00%|[605, 606, 607, 608, 1598, 1599]|
|:large_blue_circle:|com/starrocks/transaction/PublishVersionDaemon.java|1|172|00.58%|[105, 106, 107, 108, 109, 110, 112, 187, 188, 190, 325, 327, 331, 332, 333, 334, 343, 344, 345, 346, 347, 348, 349, 350, 352, 353, 356, 374, 406, 408, 409, 411, 412, 414, 417, 418, 419, 420, 422, 423, 424, 425, 428, 429, 431, 432, 433, 434, 435, 436, 437, 438, 440, 441, 442, 443, 444, 446, 447, 448, 451, 452, 454, 458, 459, 460, 463, 464, 465, 466, 467, 468, 469, 470, 471, 472, 473, 474, 475, 478, 479, 482, 483, 484, 485, 486, 489, 492, 496, 497, 498, 502, 503, 504, 506, 508, 510, 513, 514, 515, 516, 518, 519, 521, 522, 524, 525, 527, 528, 530, 531, 533, 534, 535, 536, 538, 539, 540, 541, 545, 547, 548, 550, 551, 552, 553, 554, 555, 556, 558, 560, 562, 563, 565, 566, 567, 568, 569, 570, 573, 574, 575, 577, 578, 579, 581, 582, 584, 585, 586, 588, 589, 591, 593, 594, 595, 597, 598, 599, 601, 602, 603, 610, 612, 613, 614, 615, 616, 617, 684, 685]|
|:large_blue_circle:|com/starrocks/transaction/DatabaseTransactionMgr.java|2|113|01.77%|[722, 723, 725, 726, 727, 728, 729, 730, 731, 732, 735, 736, 737, 738, 740, 741, 743, 744, 745, 748, 749, 750, 751, 752, 753, 754, 758, 759, 760, 761, 765, 766, 767, 769, 770, 771, 774, 775, 777, 778, 779, 780, 781, 785, 786, 789, 790, 791, 793, 795, 797, 1222, 1223, 1225, 1226, 1227, 1228, 1230, 1231, 1232, 1233, 1234, 1236, 1239, 1240, 1241, 1242, 1243, 1244, 1245, 1570, 1571, 1572, 1574, 1575, 1576, 1676, 1678, 1679, 1680, 1682, 1684, 1686, 1778, 1779, 1780, 1781, 1783, 1784, 1785, 1787, 1789, 1793, 1796, 1797, 1800, 1801, 1802, 1803, 1804, 1806, 1807, 1812, 1817, 1821, 1823, 1824, 1846, 1847, 1848, 1849]|
|:large_blue_circle:|com/starrocks/lake/Utils.java|3|13|23.08%|[168, 169, 170, 171, 172, 173, 196, 198, 199, 202]|
|:large_blue_circle:|com/starrocks/transaction/TransactionGraph.java|15|38|39.47%|[186, 187, 189, 190, 191, 194, 195, 197, 198, 199, 200, 203, 204, 205, 206, 207, 210, 211, 212, 213, 214, 217, 218]|
|:large_blue_circle:|com/starrocks/common/Config.java|2|2|100.00%|[]|