@wanpengfei-git 2023-08-22
#### [BE Incremental Coverage Report]

:disappointed: **fail** : 64 / 92 (69.57%)




#### file detail

|   |path|covered_line|new_line|coverage|not_covered_line_detail|
|----|----|----|----|----|----|
|:large_blue_circle:|src/service/service_be/lake_service.cpp|23|34|67.65%|[259, 260, 261, 262, 263, 287, 288, 303, 304, 309, 310]|
|:large_blue_circle:|src/storage/lake/tablet_manager.cpp|41|58|70.69%|[697, 722, 777, 781, 782, 943, 952, 953, 954, 955, 956, 957, 958, 959, 960, 961, 963]|