@wanpengfei-git 2023-08-24
#### [BE Incremental Coverage Report]

:disappointed: **fail** : 68 / 90 (75.56%)




#### file detail

|   |path|covered_line|new_line|coverage|not_covered_line_detail|
|----|----|----|----|----|----|
|:large_blue_circle:|src/storage/lake/tablet_manager.cpp|39|52|75.00%|[777, 787, 788, 866, 867, 868, 869, 970, 971, 972, 973, 974, 975]|
|:large_blue_circle:|src/service/service_be/lake_service.cpp|29|38|76.32%|[287, 288, 303, 304, 309, 310, 314, 315, 323]|